Update Hashie - requires changing namespace
