+++ spec/rash_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-describe Hashie::Rash do
+describe Hashie::Hash::Rash do
   subject {
-    Hashie::Rash.new({
+    Hashie::Hash::Rash.new({
       "varOne" => 1,
       "two" => 2,
       :three => 3,
@@ -23,7 +23,7 @@ describe Hashie::Rash do
     })
   }
 
-  it { should be_a(Hashie::Mash) }
+  it { should be_a(Hashie::Hash) }
 
   it "should create a new rash where all the keys are underscored instead of camelcased" do
     subject.var_one.should == 1
@@ -31,11 +31,11 @@ describe Hashie::Rash do
     subject.three.should == 3
     subject.var_four.should == 4
     subject.five_hump_humps.should == 5
-    subject.nested.should be_a(Hashie::Rash)
+    subject.nested.should be_a(Hashie::Hash::Rash)
     subject.nested.nested_one.should == "One"
     subject.nested.two.should == "two"
     subject.nested.nested_three.should == "three"
-    subject.nested_two.should be_a(Hashie::Rash)
+    subject.nested_two.should be_a(Hashie::Hash::Rash)
     subject.nested_two.nested_two.should == 22
     subject.nested_two.nested_three.should == 23
     subject.spaced_key.should == "When would this happen?"
@@ -64,7 +64,7 @@ describe Hashie::Rash do
 
     merged.nested.four_times.should == "a charm"
     merged.nested.fourTimes.should == "a charm"
-    merged.nested3.should be_a(Hashie::Rash)
+    merged.nested3.should be_a(Hashie::Hash::Rash)
     merged.nested3.hello_world.should == "hi"
     merged.nested3.helloWorld.should == "hi"
     merged[:nested3][:helloWorld].should == "hi"
@@ -78,7 +78,7 @@ describe Hashie::Rash do
 
     subject.nested.four_times.should == "a charm"
     subject.nested.fourTimes.should == "a charm"
-    subject.nested3.should be_a(Hashie::Rash)
+    subject.nested3.should be_a(Hashie::Hash::Rash)
     subject.nested3.hello_world.should == "hi"
     subject.nested3.helloWorld.should == "hi"
     subject[:nested3][:helloWorld].should == "hi"
@@ -92,7 +92,7 @@ describe Hashie::Rash do
 
     merged.nested.four_times.should == "work like a charm"
     merged.nested.fourTimes.should == "work like a charm"
-    merged.nested3.should be_a(Hashie::Rash)
+    merged.nested3.should be_a(Hashie::Hash::Rash)
     merged.nested3.hello_world.should == "hi"
     merged.nested3.helloWorld.should == "hi"
     merged[:nested3][:helloWorld].should == "hi"
@@ -101,7 +101,7 @@ describe Hashie::Rash do
   it "should handle assigning a new Hash and convert it to a rash" do
     subject.nested3 = {:helloWorld => "hi"}
 
-    subject.nested3.should be_a(Hashie::Rash)
+    subject.nested3.should be_a(Hashie::Hash::Rash)
     subject.nested3.hello_world.should == "hi"
     subject.nested3.helloWorld.should == "hi"
     subject[:nested3][:helloWorld].should == "hi"
